[ALPS05444004] Coverity: CERT-C Characters and Strings
CID: 10748364

MTK-Commit-Id: 3edf0de1e05a0c3987fa6eebda193fff8a9c2c93

Feature: [Android Default] Camera Application Basic Functions
CR-Id: ALPS05444004

+++ drivers/misc/mediatek/imgsensor/src/mt6768/camera_hw/regulator/regulator.c
@@ -96,22 +96,22 @@ enum IMGSENSOR_RETURN imgsensor_oc_interrupt(
 	if (enable) {
 		mdelay(5);
 		for (i = 0; i < REGULATOR_TYPE_MAX_NUM; i++) {
-			if (preg_own->pregulator[sensor_idx][i] &&
-					regulator_is_enabled(preg_own->pregulator[sensor_idx][i]) &&
-					!Is_Notify_call[sensor_idx][i]
-				) {
+			if (preg_own->pregulator[(unsigned int)sensor_idx][i] &&
+			regulator_is_enabled(preg_own->pregulator[(unsigned int)sensor_idx][i]) &&
+			!Is_Notify_call[(unsigned int)sensor_idx][i]
+			) {
 				/* oc notifier callback function */
 				if (reg_oc_debug[i].name == NULL)
 					reg_oc_debug[i].name = regulator_control[i].pregulator_type;
 				if (reg_oc_debug[i].regulator == NULL)
 					reg_oc_debug[i].regulator =
-						preg_own->pregulator[sensor_idx][i];
+						preg_own->pregulator[(unsigned int)sensor_idx][i];
 				reg_oc_debug[i].nb.notifier_call =
 					regulator_oc_notify;
 				ret = devm_regulator_register_notifier(
-					preg_own->pregulator[sensor_idx][i],
+					preg_own->pregulator[(unsigned int)sensor_idx][i],
 					&reg_oc_debug[i].nb);
-				Is_Notify_call[sensor_idx][i] = true;
+				Is_Notify_call[(unsigned int)sensor_idx][i] = true;
 
 				if (ret) {
 					pr_info(
@@ -133,15 +133,15 @@ enum IMGSENSOR_RETURN imgsensor_oc_interrupt(
 		/* Disable interrupt before power off */
 
 		for (i = 0; i < REGULATOR_TYPE_MAX_NUM; i++) {
-			if (preg_own->pregulator[sensor_idx][i] &&
-				!regulator_is_enabled(preg_own->pregulator[sensor_idx][i]) &&
-				Is_Notify_call[sensor_idx][i]
-				) {
+			if (preg_own->pregulator[(unsigned int)sensor_idx][i] &&
+			!regulator_is_enabled(preg_own->pregulator[(unsigned int)sensor_idx][i]) &&
+			Is_Notify_call[(unsigned int)sensor_idx][i]
+			) {
 				/* oc notifier callback function */
 				devm_regulator_unregister_notifier(
-					preg_own->pregulator[sensor_idx][i],
+					preg_own->pregulator[(unsigned int)sensor_idx][i],
 					&reg_oc_debug[i].nb);
-				Is_Notify_call[sensor_idx][i] = false;
+				Is_Notify_call[(unsigned int)sensor_idx][i] = false;
 				pr_info("Unregister OC notifier");
 			}
 		}
